Transformed Application Logic to MV, Observable Objects
From now on views are not responsible for holding the data, instead the application logic will have a Model which will be responsive for the data leaving the view responsible only for the view components. Using new components of the SwiftUI such as Observable Objects (Bindable Object - deprecated) to accomplish Favorite Pizza Places using a Toggle Switch, and is only responsible for a session of the app.